fix(emit): hoist es5 private storage before cjs preamble
